split calitp plugin into modules, prep for packaging
